pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sysdiagnose"
version = "0.1.0"
description = "A tool for sysdiagnose parsing and analysis"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.11"
license = { text = "EUPL-1.2" }
authors = [
  {name = "EC-DIGIT-CSIRC"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: EUPL 1.2",
    "Operating System :: OS Independent",
]
keywords = ["sysdiagnose", "analysis", "parsing", "iOS", "apple", "malware", "forensics"]
urls = { "Homepage" = "https://github.com/EC-DIGIT-CSIRC/sysdiagnose" }

dependencies = [
    "docopt==0.6.2",
    "graphviz==0.20.3",
    "tabulate==0.9.0",
    "python-dateutil==2.9.0.post0",
    "gpxpy==1.6.2",
    "pandas==2.2.2",
    "numpy==2.0.0",
    "nska-deserialize==1.5.1",
    "yara-python==4.5.1"
]

[project.scripts]
sysdiagnose = "sysdiagnose.main:main"
